// frontend/src/components/ui/sidebar/enums/sidebar.ts
export enum SidebarState {
  EXPANDED = "expanded",
  COLLAPSED = "collapsed",
}

export enum SidebarSide {
  LEFT = "left",
  RIGHT = "right",
}

export enum SidebarVariant {
  SIDEBAR = "sidebar",
  FLOATING = "floating",
  INSET = "inset",
}

export enum SidebarCollapsible {
  OFFCANVAS = "offcanvas",
  ICON = "icon",
  NONE = "none",
}